src/requirements.txt
langchain==0.3.27
langchain-community==0.3.30
langchain-core==0.3.78
langchain-groq==0.3.8
langchain-nomic==0.1.5
langchain-qdrant==0.2.1
# onnxruntime==1.19.2
langchain_unstructured==0.1.6
langgraph==0.6.8
langsmith==0.4.32
python-dotenv==1.1.1
Requests==2.32.5
streamlit==1.50.0
grandalf==0.8
sentence-transformers==4.0.2
langchain-huggingface==0.1.2